Only update cameras detail every 60s
My protect server saw a cpu spike and was a few
degerees warmer after installing the unifiprotect
integration due to the frequent polling.  When there
are >10 cameras this problem is much worse.

The bootstrap api is an expensive api to run
every 2s, which is the current update interval.

The events api is rather cheap so its not a concern.