Add global settings support for cameras (#380)
This should add full global camera settings support to the Camera
objects.

* Adds helpers for active recording settings and smart detection
settings so `use_global` does not need to be everywhere
* All `set_` helpers that edit recording settings (except
`set_privacy`), OSD settings or smart detection settings will now raise
a `BadRequest` if `use_global` is enabled for the camera. This may seem
like a breaking change, but previously, if `use_global` _was enabled_,
it would just lead to unexpected behavior because nothing would change.
* `set_privacy` automatically turns _off_ `use_global` for a camera. The
new `reenable_global` parameter to `set_privacy` can turn them back on
after disabling it.

If this is used "as in" in Home Assistant, it will lead to a lot of
exceptions if you try to use switches/controls for cameras that have
global recording settings enabled for them.