fix: Fixes new cookie flag preventing auth cookie from being stored
As of UniFi Protect 4.0.5, they have started sending the cookie flag "partitioned".
Python currently does not know about this (planned to be fixed in 3.13 https://github.com/python/cpython/issues/112713).
This causes the cookie to be considered invalid, and thus the auth cookie is never saved.
Since unifi protect rate limits authentication attempts this leads to 501 errors after a few requests are made.

This commit adds `partitioned` to the list of cookie flags the standard library accepts as valid.
Thus allowing the cookie to be correctly parsed and saved.